Update thor_connector.py
This may be a bug. When I let the model generate string-formatted actions instead of action_id, and the model generates illegal actions such as "done" and "terminate", it will directly exit here, causing the env to be unable to give the reward and error description of this action, etc. 
Even worse, due to the exception mechanism on line 150 in "embodiedbench/evaluator/eb_alfred_evaluator.py", when the action list generated by the model has only one illegal action, when running the code on line 125, the _max_episode_steps limit mechanism in the for loop on line 124 and the judgment mechanism on line 323 in "/embodiedbench/envs/eb_alfred/EBAlfEnv.py" will fail. The variable "done" will not be set to True even though the step limit has been exceeded, resulting in an infinite loop on line 85 in "embodiedbench/evaluator/eb_alfred_evaluator.py". If the large model continues to generate the same results "done", "terminate", etc., it will not be able to exit. 
btw, I lost a lot of dollars o(╥﹏╥)o